% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iEat_to_foodWeb.r
\name{iEat_to_foodWeb}
\alias{iEat_to_foodWeb}
\title{transform iEat prediction result to matrix format}
\usage{
iEat_to_foodWeb(iEatResult)
}
\arguments{
\item{Matrix, }{catalogue of empirical data used to infer predictions}
}
\value{
A dataframe with source taxa for which target predictions are made, target infered from catalogue data (empirical) and target infered from KNN algorithm
}
\description{
transform iEat prediction result to matrix format
}
\author{
David Beauchesne
}
